# openapi.yaml
openapi: 3.1.0
info:
  title: HighFive! API
  version: 2.0.0
  description: >-
    Contrat backend v2, derive de `src/domain/**` (schemas zod), `src/api/*.ts`
    (appels reels du front) et `src/mocks/handlers/**` (regles metier
    appliquees cote mock). Voir `docs/v2/backend/SPEC.md` pour le detail des
    regles metier, du modele de persistance propose et des ecarts connus.
    Les schemas de composants sont generes depuis `src/domain/index.ts` par
    `pnpm export:schemas` (zod 4, `z.toJSONSchema`) dans `docs/v2/backend/schemas/`.
servers:
  - url: /api
    description: Prefixe fixe (voir `src/api/client.ts`), au-dessus de `VITE_API_URL`.
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: Fil des projets publics actifs (R-V1)
      security: []
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 200 }
          description: Recherche texte (titre, accroche).
        - name: tags
          in: query
          style: form
          explode: false
          schema: { type: array, items: { type: string } }
          description: Filtre additif par identifiants de tags (liste separee par virgules cote HTTP).
        - name: participation
          in: query
          schema: { $ref: "#/components/schemas/Participation" }
          description: >-
            Declare cote client (`ListProjectsQuery`) mais non applique par le
            mock actuel — a implementer cote backend, voir SPEC.md.
        - name: sort
          in: query
          schema: { type: string, enum: [recent, popular, relevant] }
          description: >-
            Declare cote client mais non applique par le mock actuel (tri fixe
            sur `lastActivityAt`) — a implementer cote backend, voir SPEC.md.
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Page de projets.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedProjectSummary" }
    post:
      operationId: createProject
      tags: [projects]
      summary: Creer un projet (etat initial `draft`)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectCreateInput" }
      responses:
        "201":
          description: Projet cree, porteur = personne connectee.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Compte non `active` (suspendu).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Fiche d'un projet
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Projet complet.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "403":
          description: Projet prive et lecteur hors equipe (R-V3).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Modifier titre/accroche/description/tags/besoins/visibilite/participation
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdateInput" }
      responses:
        "200":
          description: Projet mis a jour.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Reserve porteur/co-porteur (R-PR1).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Supprimer definitivement (R-PR7, confirmation nominative)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectDeleteInput" }
      responses:
        "204":
          {
            description: "Projet supprime (differe 30 jours cote donnees, R-X3 ; voir SPEC.md).",
          }
        "400":
          description: Le titre saisi ne correspond pas (`confirmTitle`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Reserve au porteur et a l'administration.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
  /projects/{slug}/transition:
    post:
      operationId: transitionProject
      tags: [projects]
      summary: Transition d'etat explicite (publier/terminer/rouvrir/archiver/reactiver)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transition: { $ref: "#/components/schemas/ProjectTransition" }
              required: [transition]
      responses:
        "200":
          description: Projet dans son nouvel etat.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400":
          description: Transition inconnue, ou R-PR3 non satisfaite (ex. aucun tag pour publier).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Role insuffisant pour cette transition (doc 04 §3).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Transition invalide depuis l'etat courant (backend, voir SPEC.md).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
  /projects/{slug}/transfer:
    post:
      operationId: transferProjectOwnership
      tags: [projects]
      summary: Transferer la propriete (R-M2)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newOwnerId: { $ref: "#/components/schemas/Id" }
              required: [newOwnerId]
      responses:
        "200":
          description: Projet avec le nouveau porteur ; l'ancien devient co-porteur.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Reserve au porteur.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
  /me/projects:
    get:
      operationId: listMyProjects
      tags: [projects]
      summary: "Mes projets (portes, co-portes, membre)"
      responses:
        "200":
          description: Projets de la personne connectee.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ------------------------------------------------------------ Highfives
...
  /search:
    get:
      operationId: search
      tags: [search]
      summary: Recherche globale (R-R4) — projets, personnes, tags
      security: []
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 200 }
        - name: types
          in: query
          style: form
          explode: false
          schema:
            type: array
            items: { $ref: "#/components/schemas/SearchEntityType" }
          description: "Defaut si absent : `projects,users,tags`."
        - name: tags
          in: query
          style: form
          explode: false
          schema: { type: array, items: { type: string } }
        - name: sort
          in: query
          schema: { $ref: "#/components/schemas/SearchSort" }
          description: "Trie les projets uniquement. `active` = `lastActivityAt` ; `relevant`/`recent` trient par `createdAt` cote mock, faute de score de pertinence texte — a affiner cote backend."
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: "Resultats par type demande (chaque champ absent si son `type` n'est pas dans `types`)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResults" }
  /feed/discover:
    get:
      operationId: getDiscoverFeed
      tags: [search]
      summary: 'Page Decouvrir : "projet du moment" + sections thematiques courtes'
      security:
        - bearerAuth: []
        - {}
      responses:
        "200":
          description: >-
            `moment` (un projet, hors sections, jamais paginee) + `sections`
            (0 a 5, chacune 6 projets maximum, sans pagination ; `for_you` et
            `near_your_projects` sont personnalisees et absentes si vides —
            visiteur, ou compte sans interet/projet). Pas de parametres : la
            barre de themes ne filtre plus ce fil, elle navigue vers
            `/search` (voir `SearchResults`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DiscoverFeed" }
  /tags/{tagId}/projects:
    get:
      operationId: getProjectsByTag
      tags: [search]
      summary: "Page d'un theme"
      security: []
      parameters:
        - $ref: "#/components/parameters/TagIdParam"
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: Projets publics actifs portant ce tag.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedProjectSummary" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Jeton opaque (pas necessairement un JWT dans le mock : `nextId()`).
        Envoye `Authorization: Bearer <token>` par `src/api/client.ts`.
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: Curseur opaque de page (issu de `nextCursor`, `null` = derniere page).
    LimitParam:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100 }
      description: Taille de page (20 par defaut cote serveur).
    SlugParam:
      name: slug
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Slug" }
...
    TagIdParam:
      name: tagId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: "`validation_error` — corps ou parametres invalides."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    Unauthorized:
      description: "`unauthorized` — jeton absent ou invalide."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
...
    NotFound:
      description: "`not_found` — ressource introuvable."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
...
    UnprocessableEntity:
      description: "`validation_error` — corps syntaxiquement valide mais violant une regle metier (ex. R-PR1)."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
  schemas:
...
    ApiErrorBody: { $ref: "./schemas/ApiErrorBody.json" }
...
    DiscoverFeed: { $ref: "./schemas/DiscoverFeed.json" }
...
    Id: { $ref: "./schemas/Id.json" }
...
    PaginatedProjectSummary: { $ref: "./schemas/PaginatedProjectSummary.json" }
...
    Participation: { $ref: "./schemas/Participation.json" }
...
    Project: { $ref: "./schemas/Project.json" }
    ProjectCreateInput: { $ref: "./schemas/ProjectCreateInput.json" }
...
    ProjectSummary: { $ref: "./schemas/ProjectSummary.json" }
    ProjectTransition: { $ref: "./schemas/ProjectTransition.json" }
    ProjectUpdateInput: { $ref: "./schemas/ProjectUpdateInput.json" }
...
    SearchEntityType: { $ref: "./schemas/SearchEntityType.json" }
...
    SearchResults: { $ref: "./schemas/SearchResults.json" }
    SearchSort: { $ref: "./schemas/SearchSort.json" }
...
    Slug: { $ref: "./schemas/Slug.json" }
...
    ProjectDeleteInput:
      description: "R-PR7 — confirmation nominative (schema anonyme, src/api/projects.ts)."
      type: object
      properties:
        confirmTitle:
          type: string
          description: Doit correspondre exactement au titre courant du projet.
      required: [confirmTitle]
      additionalProperties: false
...